pyproject.toml
[project]
name = "tmr-esports-modmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "sqlalchemy>=2.0.36",
    "discord-py",
    "rich>=13.9.2",
]

# [tool.pyright]
# reportMissingImports = true
# reportMissingTypeStubs = false
# reportImportCycles = false
# reportPrivateUsage = false

[tool.uv.sources]
discord-py = { git = "https://github.com/Rapptz/discord.py" }
